DataRange implemented, and added Mongo to makefile
DataRange allows to defer loading various data chunks, and allows combining various storage types into a seemingly seamless range of values. There are some annoyances: RangeList requires no two datapoints in a stream to have the same timestamp. This is annoying, but it is a fundamental design decision, so it just has to be.