TimeBatchDB is functional
The database tests pass. It is still not going to be fast, since I short-circuited through half my code in order to get it functioning enough to move to the next step.
In particular, it dumps data straight to MongoDB, and completely avoids the appendfile and Redis caching. Redis caching in particular is not yet ready.
Those things can be completed at my convenience now, since the database inputs and outputs are identical to the future version (just not as fast).