Rewrote type core
I realized that there is a much simpler way to do types, but which required a rewrite. The type core is now equally functional to the old version. It furthermore allows array types for binary and string data - so a string with at most 12 characters is of type text[-12] (negative means "at most").

The new code is much simpler, and I moved it all to the dtypes subdirectory - the type core is separate from the database itself.
Types now include:
binary (optional [] as length check)
text (optional [] as length check)
int
float
bool

Relevant to issue #13