Android: DataCache
DataCache is the sqlite database which holds all data before it is syncd with server. GPS is logging to it correctly now (although it is only logging lat+long atm, none of the accuracies and other stuff is in the data array).

This marks the moment where the android app can start being useful - it is saving data. One connecting with server is implemented, it can dump this data.