TimebatchDB refactor part 2
Deleted all code that isn't strictly necessary for the database in datapoint/datarange/datapointarray, increased test coverage to 93% average in files being committed.

Also wrote basic benchmarks of the most critical tasks.

The files being committed right now form the underlying core of the database. This commit leaves the database in a broken state as the higher level stuff was not yet refactored.