TimebatchDB: mongo->sql conversion
The mongoDB storage method was removed from the database (if we ever need it, we can just resurrect it from git history).

The database now uses SQLite3 for its storage mechanism. Once postgres support and config files are committed, an initializer for Postgres can be easily added.

Benchmarks show rather poor performance of Sqlite in inserts of data - about 0.1s per insert. They show ~0.02 per get query. Hopefully this will not be the case with postgres.