TimebatchDB: refactor part 3
Rewrote the RedisCache to be simpler (it no longer has its own RedisRange). Completely rewrote tests, and wrote several extra benchmarks.

Performance for RedisCache is pretty good. I was abole to get ~10k inserts per second into redis on the VM on my laptop. This is definitely a nice thing.

Also deleted the old keymap code. It is not necessary anymore (coldstore will have to wait a while, storing data in postgres will do for the forseeable future).