TimebatchDB: Performance tweaks
Turns out the low read performance was entirely due to sql compression. I have disabled compressino for sql, and I get 620 reads of 1000 datapoints per second. This amounts to 620k datapoints read per second. This will have to be explored more with postgres, especially if we go with larger batch sizes (1k or so), since the compression is less of a performance issue the larger the batch.

I have also profiled insert performance, and almost doubled the speed by minimizing redis calls. Now it can do ~4k inserts of 10 datapoints each, which amounts to about 40k datapoints/second. If inserts are one by one, it would be much slower, since I also check for timestamp validity.

It looks like for now batches of ~100 datapoints are a good place to start.