Fixed some of my mistakes last night plus DRY and unittest upgrade.
I removed the guard code over the comitter, I now see why it was done.

I modified the test suite to kill postgres and remove the database so
we stop erroring out when a database is used locally.

I also added guards to the test suite to not allow it to be run as
root.

I modified userdb to include the sqltype as @dkumor did rather than
storing it in an internal field so we don't have two of them hanging
around, this can all be pushed up in v2 from the userdb into unified.