Schema: Initial commit
Dtypes was completely scrapped. Using JsonSchema allows for a much more powerful data storage in streams. And it was all of 50 lines of code.

I am currently using JSON as the internal storage mechanism. This is because using BSON does not allow simple (one line) encoding of single-valued streams. JSON does, and as such is what is used as the internal byte storage type of Schema. If it ends up being inefficient enoguh to warrant attention, I made the structure readily extensible to hand-tailored binary schemas.