StreamDB: DatapointReader implementation
The DatapointReader interface is a generalized iterator. RangeReader implements the interface, and given a timebatchdb DataRange and the schema to use, returns the datapoints in the correct type.

JsonReader takes a DatapointReader and implements the io.Reader interface, allowing the use of io.Copy to stream data to a user. It outputs a json array of the datapoints.

In the future, will probably want to implement csv and xml variants of this.