Android: Google fit
Gathers data from google fit. I am starting to think that google fit would be better as a python app which uses oauth to get the user data... Java is just so much of a pain in the butt to use. And we are assuming that the user only has one cellphone with DataConnect.