ConnectorDB: Broke it into many files
Also fixed starting and stopping so that we are not dependent on unix tools. We no longer need to kill stuff, the database makes sure servers are killed.

I now have settings loading from config files that are rewritten on each run with new settings. This allwos us to set ports on each run.
I also had pid and log files set in .conf files for the individual servers. So we don't worry about those

I temporarily removed loading from cdb.ini config file, since some stuff changed there. We now need port numbers of the different servers, and the default interface to look on

I broke the code into streamdb/dbmaker

For some reason test of users fails. I am not sure if that is due to the modified connectordb.

Lastly, I temporarily removed the capability of running the rest server or web interface from connectordb. In particular, they will require actually writing the pid file and checking if the process exists on stop. I can implement that it a couple moments, but didn't get around to it yet.

I think that this will allow tests to pass once the weird users failre is fixed.