StreamDB: Added LRU cache
The cache allows us not to thrash postgres each time a user makes a request. But that was really a byproduct.
The real reason for the cache is that it makes coding devices so much easier - rather than being very careful about efficiency and passing devices around each function, the code can be greatly simplified by simply getting the device each time, since it will be in the cache.
This allowed me to fix the Update* functions to not need passing in of unmodified stuff - you just pass in the name, and all works fast and well.

This also automatically makes websockets work great, since the cached stuff expires after 1 minute, so we get updating in-use operators as a bonus.