StreamDB: LRU Benchmark
Without the LRU cache on, we can get about 567 logins/s
With the cache on, we get 179,500.

This speed difference would be further compounded when running update/create commands on own user - I just tested straightup login.

One thing to note is that it would be possible to make nocache go to ~1000/s if the user and device were querie in one query.

ut as it stands, there is no need for that! With the default settings, the LRU cache will allow HELLA speed.

Real-world speed will be somewhere in between the two - nocache is worst-case scenario where each user is new, while cache is the best-case, where there are a consistent ~100 users (that is the size of the cache right now - it can be set to anything,though)