Fixed Web service CWD bug
This bug breaks tests, and exists in master branch. It looks like the test script is broken on master, since the tests fail, but the test script still returns success. I had fixed this bug a while back in REST.

As it stands, I don't know if the fix makes web work, or breaks it. But at least tests now pass (current directory is not changed to the executable directory).