StreamDB: Initial Restructure
Got rid of a lot of code duplication. Right now, most testing is done in streamdb, and not in the individual modules, since current tests would require import loops.

There is now a BaseOperator interface that things conform to, and the BaseOperator is extended to a full Operator which has the duplicate code.