ComputerLogger: Initial syncing
The code for computerlogger is more of a proof of concept - it is all over the place, and will probably need to be rewritten.

BUT, the current code has a proof-of-concept computer logger which works and syncs data with a server using the python logger.

One thing that is showing up as something annoying is actualyl creating devices: currently we require that this is done manually, but there should be a better way.