Workaround for js jsonschema issue
The jsonscema library thought it was serverside, so was replacing _ with require("underscore"); Commenting it out works aronud the problem.

Also fixed a further issue to make the form actually display (schema was a string, and the js lib for jsonschema takes a dict of schemas rather than one schema).