Javascript: Initial client tested
Running tests is as simple as opening connectordb_test.html while a database is running on port 8000 with admin user test:test.

Some changes were required in REST to make cross-origin requests work for testing (and for the library to work anywhere when on browser side).

Datapoint inserts/gets are also now supported.

In the future, websockets and subscribing to inputs should also be enabled.