Security update: hide 500 errors and a builder for the whole thing
This commit creates a standard way to apply http wrappers that modify
security content of the site. Logging has been tested, same for 500
and the security headers.

Additionally, the "run" command now uses the wrappers by default, and
the web client has two special URLs for generating 500 and 400 errors
for external testing and validation.