Android: Star ratings
Basic star ratings are working - they are the basic things which will be
the feedback of the user. They need to be a bit more responsive, but
everythnig seems to be basically functional.

All that needs to happen now is sync, and background timer sync, and the
app is ready for testing!