Android: Added depth and motivation ratings
Motivation is self-explanatory, it is how pumped you are for life in
general. It is different than productivity, because you can be productive
while not feeling motivated.

Depth on the other hand is due to the current ratings not being really
adequate at quantifying actual emotion. A bittersweet memory, or
contemplating the universe does not
make you feel good nor bad, it does not make you productive, nor does it
bring you anywhere close to your goals. But it adds a certain depth to
life. So 0 stars for depth is "I am a robot, doing what needs to be done".
So the more "life" in your life, the more stars.