Datastream: Have SQL database running
The sql database was changed to reflect substreams and the fact that timestamps are now floats.

In other news... I benchmarked postgres - and as expected, it is slow as balls on individual inserts... but once we get to reads, thanks to the batching mechanism, we can go REALLY fast - even faster than redis thanks to the batching mechanism. In fact, my benchmarks show that we can optimistically read ~600k datapoints a second from a stream through postgres (if in memory). This is an absolutely incredible result.