Easier SQL modification/debugging.
Commands in the shell are now sorted by name, so help will return more useful
results.

The postgres command no longer does a fast stop, which was triggering problems
with the runtests on my system as postgres hadn't fully shut down before the
next test was started.

On a database upgrade failure, the conversionscript is now printed out so it
can be checked for syntax errors.

The shell now got a SQL command, which launches a psql shell within the
existing session to do modification/lookups on the running database. This will
make things easier with named databases, remote servers and the like.