Fixed 1:1 mapping issue so we don't fail primary key constraints
The 1:1 function I had us set up with before wasn't correct,
this new one uses a PRNG that has a known cycle through the
keyspace and fits a congruency so that we know it is 1:1.

A brute force check over the entire 2^32 space shows that the
generator indeed is a 1:1.