Made log display prettier.
Now, if there is an error, the debug message along with timing information are upgraded to info level - so we have more info than just the error message. Also made the debug message passed on, allowing to not use extra lines of logging in debug mode to write query times.