TimePlusIndexRange
This function is necessary in order to be able to do Datasets correctly, since several interpolation methods actually need a couple datapoints *before* a given time.

The current implementation is as simple as they go, meaning that it usually uses 2 queries - one to get index at timestamp, and another to get the new modified index datarange. This can be avoided if a bit of refactoring is done, as batches oftentimes include datapoints before them when queried.