Started interpolator implementation
It took me a couple days to figure out how to implement this in a step-by-step fashion.

This is the start of the "initial" dataset implementation. The goal now is to get everything functional and giving correct results. The implmentations are therefore stupid-simple, albeit not very efficient.

After the basics are all completed and testing framework is in place, cleverness can be added (substream-based aggregations, smart skipping of ranges, and so forth) to avoid having to read through the entire stream in many very common cases.

There is a LOT of improvement possible, so this stupid-simple implementation nevertheless retains some forward-thinking, and makes some seemingly-weird decisions, like passing the full database to interpolators rather than just the associated dataranges. This will allow interpolators to be clever on asymmetric streams, to create "aggregation indexes" as substreams, etc.

It is funny how hard it is to decouple the complexity of what you want something to do from the absolute basic underlying structure that will allow iterative buildup.