Changed DataRange to StreamDataRange
StreamDataRange was most useful within datastream for manipulatino of large ranges of datapoints and getting metadata.

Now that we are doing data analysis, we want a simple interface which just includes the iterator portion of StreamDataRange, which is the new "DataRange".
A StreamDataRange is a valid DataRange, so nothing really changed.