Changed StreamDataRange -> ExtendedDataRange
The new naming is more consistent with what it represents - an extended version of the DataRange iterator that is returned by the operator functions.

When casting, an ok check must be placed, since not every datarange is an ExtendedDataRange - datasets and merges are not ExtenededDataRanges.

Transforms are - but their indexing is wonky.