REST and web combined into one
Now both rest and web are consistent with each other. I liked the template based code that @josephlewis42 wrote for the app, so I extended it. Now the templates are automatically reloaded on modification. Furthermore, the cookies that allow login also enable rest api, so that auth doesn't need to be handled from js.

404 handlers are now more clever (giving pretty page based upon location).

The entire login works a bit like github now:
/ is index if not logged in, and the app index (profile) if logged in.
/user and device/stream give login pages if not logged in, and show user/device/stream pages if logged in based upon permissions.

Now all that is left is to finish the login api itself, and we're good to go!