Auto-Reload configuration
The configuration now reloads itself automatically upon changes to the configuration file. This allows changing properties of the server without restarting.

Most properties can't be changed, and will just fail silently, but most things in the frontend are runtime-modifiable.

This will be extremely useful in changing join properties while the server is running (such as max user number, etc).