Join: Complete
Joining now works correctly.

I would like to reiterate that I think that our current DatabaseOperators are MUCH too confusing. It took me a good chunk of time to figure out which operators are used where.

The join does not validate emails - this is basically a minimal proof of concept. It assumes that star ratings start with "rating_" and goals start with "goal_".