Read/Write User with permissions
I ended up switching to a tag based system, which is more robust.

In order for permissions not to leak information, we needed to switch modifications to map[string]interface{}, since unmarshalling into the struct did not tell us if a value was attempted change (and any way to do it would give info leaks).

Reading was also done with maps to make it clear which permissions your device has. That is, only the fields for which your device has permissions are given.