Structured User/Device/Stream Creation
The creation functions were starting to get unweildy, especially since we want to
be able to do more advanced operatinos during creation, so I put all of the arguments in a single struct for creation.

Also, while it is not implemented as a single transaction at the moment, this work lays the foundation of generating a full device or user, with all of its streams right away in one transaction. This will be directly exposed to REST, so while it won't be particularly fast right now since creating all the devices and streams happens in their own queries, it allows us to optimize it all into one transaction in the future.