Middle of refactor
I was unhappy with how redux was being used (read: basically not at all), and didn't like the massive single file of User code.

I spent most of the night thinking about it, and I think I figured it out. Unfortunately, I am too tired to finish implementing the changes today, so committing what I have.

This commit is broken - but once it is fixed, I think it will work well.

In essence I am splitting up the redux store - all user/device/stream data will be held in indexedDB, and kept synchronized with ConnectorDB outside of redux. The storage API works similarly to redux (the components subscribe to changes). So while app data isn't held in redux, app state _is_ - including per-user page states, which allows the app never to lose state - so long as it is not refreshed (I don't know if I want to keep state past refresh, since it invites "broken app" on bugs which can't be worked around with a refresh.)