Added ability to query list.
localForage has the weakness that it is extremely limited in what can be done. I need to be able to query by child devices, etc, but that is not easily done. For now I have hacked a solution using localforage-startsWith, but that won't work for streams, since they are another level of query.

I'll figure out how to work around that tomorrow...