Cobra CLI
The old cli code was very convoluted, and made starting with ConnectorDB's code much more difficult than it had to be.
The new CLI code is nicely modular, and only one flag (in shell) is different.

ConnectorDB now returns error (instead of 0) when commands fail, which allows simple checking for success.

Closes #298