Preliminary version of daemonized start
connectordb start database now starts the frontend too. This still needs work, as no flags are transferred to the new process, and you can't just start the backend only. But this is the proof of concept - the rest are fixes.

Oh, and also - while wondering how to implement config transfer between processes, I added logfile and loglevel to the config - now logging is live-configurable :D