Fixed data update bugs
The data update mechnaism wasn't working properly - it was updating with wrong properties. Also, the data does not need to be set in state, since it is only updated on prop changes, which force updates anyways.

Lastly, the chartjs component is refreshing for some reason after querying data. I'll figure it out later.