Added Redux-Saga and made Downlink & Analysis page
When making the android app, I found redux-saga extremely useful.
It allowed me to change the entire structure into one that is extremely
simple to comprehend and extend. The ConnectorDB frontend was built
as I was learning React, and as such it has many issues with where actions
happen and how components are rendered, as well as how the store is structured.

All further code from this moment will conform to the simpler, more extensible format. Although it makes the app temporarily a hodge-podge of methods, at some point the old code will be refactored to use the new functional/saga format.

The reason for this shift was because I wanted to add the downlink and analysis pages, and the old code structure made me dread the amount of work it would take to do so. This meant that the old code was bad, and something needed to change. So I basically tacked on saga on top of the old codebase, and am developing the new parts with it.

The downlink and analysis pages do not yet have any functionality. Everything is linked together in the backend, though, so the codebase is ready to move forward.