Analysis Page functional
This means that everything seems to be working correctly. More testing is needed to make sure nothing weird is going on. Also, there are no visualizations that work for datasets. The next step is adding a couple that can support them, such as a scatter plot and multiple time series to the line plot.

And lastly, I'd like to add color-based map plotting.