0.3.0 Feature-Complete. Debugging Time!
All features of the database and its frontend are implemented. It is now time to clean up the code, and find the bugs that showed up. Since analysis is functional, the dataset code needs more testing (several issues were notcied during frontend writing, such as a programming error in dataset). The biggest worry is timestamp inconsistency in export. I will need to find out where it came from and fix it - even if it means rewriting the entire data storage backend...