Started work on 0.4
The main issue I found with 0.3 was that it was a one-trick-pony. It could gather data, and display a couple plots. Any analysis needed to be done outside the app itself, and no integrations could be added without great difficulty.

To fix this, the new version is going to be built upon a powerful plugin system, which will allow modifying and adding features - in python!

The new version is also going to get rid of the complex setup procedure, being distributed as a single executable.

Since these changes affect the core of the database, the base needs to be rewritten. This is also an opportunity to simplify it greatly - no more batching or complex data insertion - I have learned that making the CRUD stupidly simple is worth so much more than having it be fast, simply because most people don't need it to be fast, but they do want to have easy addition of features, and deleting datapoints, etc.

I expect this version to take most of the year to complete, so this is the start of a basic plugin system.

The current code doesn't do anything yet - it is a basic structure - the plugin overlay filesystem, packed assets, and a configuration merging code. I will need for HCL2 API to become more robust before the config can be expanded.

Nevertheless, it is time to use the lessons from 0.3 to create a powerful, extensible, and simpler 0.4.