Have a working app javascript splitting mechanism
The app is no longer a huge monolithic bundle, but rather a conglomeration of a bunch of separate javascript files. This has the benefit of loading less at each import, and also allows overriding the files with plugins.

In the process of getting this working, went throguh webpack and separate folders for each javascript file (ie separate webpack setups). After a lot of frustration, finally settled on direct es6 with rollup, which gives me everything that I want, including jsm modules for the components, which allows dynamic loading.