Further skeleton setup: grpc,rest,docs, and setup
A lot of code is going to be auto-generated using tools in the new version of ConnectorDB. In particular, the main method of communication with the server is going to be grpc, with the REST API layered on top of grpc. This will happen transparently to the user, but it will allow auto-generated libraries for multiple languages.

Using protobuf 3 also allows auto-gen of docs for openapi.

Finally, moved the setup portion of cdb to rollup, for consistency across all code.