Got grpc with REST working...
Ooof. That was rough. The setup has the hard requirement that CDB always use https.
This means that it will need to auto-generate certs, and do some serious magic to make it as transparent to the user as possible.

The worst part is the default browser failure for self-signed certificates - they made it really hard to serve https on localhost...

Still needs a bunch of work with certs to figure out what the hell is going on with them...