Started registry work, half finished
The registry uses the gituhb API to extract a plugin from a github repository.
The extract from github code is in a state of half finished, meaning that all the stuff I wasn't sure how to do
is working, now it just needs validation and linking together with the registry itself.

However, now that the hard part here is solved, time to move on to the server.