Yet another database refactor
The scopes were set up in a messed up way, and the sql was all over the place. By not worrying about efficiency, and putting all the reusable logic into database views, the code is now much cleaner.

There is also the question of how efficient this new code is, since technically, with a lot of users, the user_scopes view might require an enormous amount of computation to be generated, since it is literally 3 tables joined together. Hopefully it won't be a real issue.

With this, I can now move forward with streams.