Removed user scopes & changed to sources
User scopes didn't give much useful flexibility, and added a huge amount of complexity to the database.

Sources are a huge upgrade to streams - instead of having just streams, the database knows about "sources", with a built in "stream" source type. Sources allow for simple extensions to the database: an aggregate stream can just be a different source, and a jupyter notebook can likewise be a source.

This required a big shift in the database, so to manage it, simplifying scopes was really necessary.