Groups & Streams are now sources, and plugins
With the concept shift to data sources, it became clear that both streams and groups can fit into the same paradigm. This actually greatly simplifies the "core" code, since there is no special handling for streams within groups.

This change is not yet complete - to enable same functionality, must implement source sharing with other users, public, and specific people. So the sharing functionality that was originally to be based on groups is now moved up to sources. It is not yet coded up.

Also, the decision was made to develop the streams and groups as completely separate plugins, and a method to integrate them with core will be developed closer to release. The reason for this, is that this will ensure that the API I develop for plugins will need to be powerful enough to handle all the features of streams and groups! It will also enable testing the plugin architecture thoroughly, and dev updates during server runtime, so the groups and streams plugins can have published "dev" versions that can auto-update to most recent versions automatically on users' servers.

However, it is important to note that the plugins will be compiled into heedy for release, so they will have a dual role - separate executable plugins for development, builtin for release.