Source routing basics working
Not yet fully happy with it, since source paths end with / here, and don't end with / anywhere else, and it is non-trivial to handle people going to the url for a source not of their type. Solving this will require a bit more thought, which can be left for later.