Removed fullname from sources and connections
fullname was one of the remainders of the ideas from CDB, but in reality, it is not needed in heedy - name is actually just a noremal name, we don't need url-based access like on github. Therefore, no more distinction between a programmatic name and a full name. name is now a fullname. With this change, the user gained a "username" element. The backend looks like it is functioning with the changes, and the main issues with the frontend are also fixed, but the creators for sources and connection still need to be updated.