u/c/s live updates & avoid api calls on websocket
The frontend is now fully live, meaning that it shows updates as they happen in users/connections/sources. This also means that it doesn't perform API calls unless they are necessary, since no need to query the database if stuff is being kept up-to-date!

Not only does this look pretty good to have heedy update to changes the moment they happen (add a source from another computer? it shows up automatically!), but it has a chance of actually *lowering* the number of API calls, since the frontend no longer queries for updates on navigation to previously-visited pages.