BROKEN: Passed runtypes through configuration
The configuration was upgraded so that the new runtypes work. I also took the chance to fix plugin settings, as well as add verbs to the routers in configuration. Unfortunately, since this is part of a pretty large rewrite of exec/run functionality, this commit leaves heedy in a broken state - the configuration is upgraded, but the downstream code doesn't yet understand the new config.