Added 'dismissible' to notifications & bugfixes
This commit is focused mainly on improving notifications to make them more useful for plugins.  This includes making notificatino actions look good, and allowing certain notifications to not be dismissible by the user. That enables things like the fitbit app to show a notification with instructions until the user actually follows those instructions to set up the app!