Added initial timeline view
The view is currently quite hacky, but I am already extremely impressed with how great it looks. Right now I'm adding the basic views as hack-jobs just to have some form of visualization, and once the view API is stabilized a bit, I'll go back to make it more robust to different data types and errors.