Added timeseries day-view
This view really helps in seeing patterns that only show up in large amounts of data. However, in doing so, I have discovered a pretty critical bug in durations: duration overlaps. Should they be allowed? Probably not in a single time series. The heedy code itself needs to check this and depending on settings, either auto fix, or rejest updated data.