td -> dt and /api/heedy/v1 -> /api
The idea of a /heedy/v1 is unnecessary: if there is ever an api v2, can always just put it at /api/2. And the heedy split is likewise unnecessary, since it gives stuff in the heedy namespace "special" status, whereas it should just be an API.

Also, td to dt makes it clear that "dt" is a delta t, ie, a duration for a datapoint.