Timeseries overlap check
The data in a timeseries is not allowed to overlap. Triggers were added
to ensure that this holds at all times.

This commit simply makes the problem result in an error, and makes the
database maintain consistency. It does not solve the issue, since an
actual solution would require quite a bit more thought on how to combine
data where durations might overlap. The full solution will have to wait
until a good approach comes to mind.