fixed duration overlaps in update insert method.
The update insert method is the equivalent of a "just works" way to insert and update data. It needs to handle issues gracefully, such as existing datapoints with overlapping durations. The current implementation deletes all datapoints which start during a new datapoint's duration, and truncates the durations of datapoints that start before the currently inserted datapoint, but overlap it with their duration.

This adds 2 queries per datapoint, but it isn't an issue, since the insert/append methods can be used by people who want higher speed, and who are willing to sacrifice ease-of-use for speed.