Simplified event router & config "on"
The event router was compeltely rewritten to be stupidly simple, since the old one was buggy. The configuration event system was also upgraded, to allow both multiple handlers for the same event type, and to allow constraints on event handlers, such as object type. This is needed to allow plugins implementing object types to only subscribe to events relevant to that type!