Upgrade frontend deps & converted table visualization
There seems to be a bug in rollup-plugin-commonjs that breaks the frontend, but don't currently have time to write a good bug report for it.I simply upgraded to v14 instead of v15, which works fine. Similarly, rollup-plugin-node-resolve still is stuck on the old version.

In dataset visualizations, insert code was moved out and is explicitly added to the timeseries rather than as a visualization. Also, the old datatable code was converted to the new system, and now seems to function correctly. The code was upgraded a tiny bit, to allow displaying object keys, but further upgrades will be needed to make it work well for datasets (like multi-series support, and named columns from schema).