Added CPU profiling and parallelized timeseries
So it turns out that go's sql interface is actually super slow, and profiling shows that inserts and queries are doing pretty poorly. This leads me to the conclusion that batched timeseries are actually something I want - and that the data should be gzipped in the database, to minimize its size. This greatly adds to the complexity (basically halfway to connectordb), but I think it can lead to huge performance improvements, and greatly decrease database size, so might be worth it.