New timeseries backend
So originally, heedy's timeseries were in the "standard" format, with a row per datapoint.
However, this led to the following problems:
1) Extremely slow inserts - 1 million datapoints took over 40 seconds!
2) Slow reads - a query on 1 million datapoints took 5 seconds, and profiled it was spending a huge amount of time in Scan
3) Very large disk footprint - a database with 1 million datapoints was over 100MB, whereas the pure json array was ~30MB

The slow reads are largely due to problems with golang's Scan being slow.
Slow writes were due to a large amount of checks that needed to be done for overlapping datapoints and such. Finally, a lack
of compression support was the culprit of large space usage.

I therefore chose to rewrite the timeseries table by batching elements. This is not a decision I take lightly - it is generally bad practice
to ""roll your own" storage format within SQL, and all interpretability is lost for timeseries rows, so they now look like gibberish, and
they can no longer be analyzed by SQL json primitives. Not to mention how difficult and bug-prone reading/writing this custom format can be.

On the other hand, the benefits are fixes to the above 3 issues. Reads are now extremely fast, each batch is compressed, lowering database size,
and large inserts happen much faster. The usage of the timeseries in heedy works well with a batching mechanism, since PipeScript is used to filter
the data rather than raw SQL. I therefore deemed it more important that heedy can query and add to timeseries extremely fast than the simplicity
of the direct approach.