CRITICAL: fixed query indexing bug
Fixed a critical error in indexing data (wrong value was being queried).

Also made schema validation go outside txn:

The schema validation actually takes >2/3 time of the insert. By having it done before the actual insert, the database is locked for a much shorter time.

Originally, 1M insert took 6s, now it still takes 6s, but only 2s of that are in a transaction - and even that is almost entirely compression of batches on append.

As of now, the main component in both read and write from tiemseries is compression/decompression, and that is not something easy to change. Overall, timeseries are now fast enough for heedy's usage, so it is time to focus on moving forward with the frontend/dashboards/visualizations.