Switched timeseries backend from json to msgpack
So this change makes compression a tiny bit worse, but it greatly decreases the processing time for encoding/decoding batches, which makes inserts go much faster. It will also greatly speed up querying on single core systems or systems with multiple concurrent queries.

The efficiency of heedy's timeseries backend has undergone a massive shift overall.
Database Size: 120MB -> 5MB
1M Query: 6s -> 1s
1M Insert: 40s -> 1s

What's funny is that the architecture here is basically identical to ConnectorDB, with the difference that there is now a bunch of code enabling inserts,deletes and modifies instead of just appends, and we no longer have a redis cache, since single-datapoint inserts are not as common as I thought - and honestly modifying a single batch is plenty fast.