Caching analysis results & initial histogram vis
There is now code that allows performing queries on the data arrays with cached results, so that for example, a min of the dataset only needs to be computed once. This allows easier use in analysis functions, since now they can just compute things when they need them independently.

Also added a basic histogram visualization. Still needs some work for categorical data and for transform results, but works nicely for numerical binning.