Added timeseries types
The types allow complex schemas to be used with timeseries, and allow menus to be shown to users so that they don't have to edit the schemas.

This completes the features that were previously present in ConnectorDB, but were not implemented in heedy's first alpha, other than manual csv import,m which will come at some point.

With this finished, it is time to upgrade the plugins to the new version, create a couple new plugins (while fixing bugs), and then release alpha 2.