Fixed superfluous user query
The frontend was querying the user on startup despite the user coming in directly in appinfo payload. This fix allows a cached user to be used even if websocket is not active if the user was queried within the past second.

Also added downsample to preload, since timeseries worker loads it.