Made updates to plugins/config replace process
To allow config updates or adding plugins, the heedy server needs to restart.
Previously, the server would shut down, and start a new instance in a background process.

This led to a couple annoyances, like a server started with heedy run going into the background on update, and issues with docker thinking heedy was finished when it was just restarting (#342). This fixes the issue on unix systems, by replacing the current process with the new one, which inherits the current pid.

The old behavior remains on windows, but heedy doesn't support windows yet, and windows isn't really used for server scripts.